test(ai): split conquest wiring test to satisfy 1000-line repo-lint gate (Refs #2924)
phase_planner_conquest_wiring_test.dart exceeded the 1000 non-comment-line
repo-lint limit (1081 lines) after the Path E feasibility-gate tests landed,
failing the Quality check. Extract the stalled-expansion own-territory
frontier-march group into phase_planner_conquest_frontier_march_test.dart;
behavior and coverage are unchanged.
